refactor(filesystem): extract local.rs's mount-registry layer
FIX 1 and FIX 2 added enough logic to local.rs's mount-registration area
(ensure_scoped_mount's wiring, the dynamic-mount LRU bound) that it earned
its own file, mirroring the existing local/fd_resolve.rs extraction.

Moved to local/mount_registry.rs: mount_local, mount_local_per_leaf,
resolve_mount_target, the LocalMount/MountTarget types, and the actual
ensure_scoped_mount logic (renamed to ensure_scoped_mount_dynamic, an
inherent method — Rust allows an inherent impl to be split across files,
but not a single trait impl, so the RootFilesystem trait impl itself,
including its one-line ensure_scoped_mount delegation, stays in local.rs).

Also:
- Registered mount_registry.rs in the fd-rooted architecture gate
  (reborn_local_filesystem_fd_rooted_gate.rs) alongside local.rs/
  fd_resolve.rs, and added a proof-of-binding test that plants a
  tokio::fs violation in it and confirms the gate still catches it there.
- LocalMount needed pub(super) visibility to stay nameable from local.rs
  once it moved to a child module, which newly exposed it to the
  deployment-mode typename ratchet (it was always named LocalMount, just
  previously private and invisible to that scanner). Added it to
  FROZEN_OTHER_MODE_TYPES as a justified Bucket-3 name: "local" here means
  "on the local filesystem" (matching local.rs/DiskFilesystem's own
  naming), not the LocalDev/hosted-tier deployment-mode axis that ratchet
  polices.
- cargo fmt cleanup carried along on a few FIX 1 lines that weren't
  fmt-clean.
